requirements.txt
diskcache
modern-bencode
simple-dht-node
